[Fix #14] Use :default-fn for :dir command line option
Because of the way tools.cli threads default values, the default dir was always
conjoined in the :dir key. This is not what we want, cause if the user
specifies the --dir option, she likely does not have "test" in her code base.
This patch uses the new :default-fn feature (present in tools.cli 0.4.1) to set
the default value *after* the command line option parsing.